% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{import_counts}
\alias{import_counts}
\title{Import the counts-based RNA-Seq dataset}
\usage{
import_counts(input, abr_healthy, abr_case)
}
\arguments{
\item{input}{Character. Path to a folder containing
\itemize{
\item \code{counts.txt}: gene × sample matrix
\item \code{gr.txt}: one group/condition per line
}}

\item{abr_healthy}{abbreviation for healthy group; for example H for healthy}

\item{abr_case}{abbreviation for case group; for example D for disease}
}
\value{
A \code{dds}: a \code{DESeqDataSet} object
}
\description{
Reads in a plain counts table plus metadata (or both via a \code{list}), create dds object,
}
\details{
Internally this function:
\enumerate{
\item Reads the two input files via \code{read.table()}: counts.txt and gr.txt or list of these two
\item Constructs a \code{DESeqDataSet} and return \code{dds}
}
}
\examples{
\dontrun{
pipeline_counts(
  input = "inst/extdata",
  outdir   = "counts_output"
)
}
}
