% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_salmon.R
\name{import_salmon}
\alias{import_salmon}
\title{Run the Salmon-based RNA-Seq pipeline for the gene-of-interest}
\usage{
import_salmon(input, abr_healthy, abr_case, tx2gene)
}
\arguments{
\item{input}{Character vector of paths to Salmon output directories.}

\item{abr_healthy}{abbreviation for healthy group; for example H for healthy}

\item{abr_case}{abbreviation for case group; for example D for disease}

\item{tx2gene}{Path to an \code{.rds} file or data.frame mapping transcripts -> genes.}
}
\value{
A \code{dds}: a \code{DESeqDataSet} object
}
\description{
Uses Salmon quantifications plus a tx2gene mapping to build counts,
}
\details{
Internally this function:
\enumerate{
\item Imports counts via \code{tximport::tximport()} and the provided \code{tx2gene} mapping
\item Constructs a \code{DESeqDataSet} and return \code{dds}
}
}
\examples{
\dontrun{
pipeline_salmon(
  salmonDirs = list.dirs("inst/extdata/salmon", recursive = FALSE),
  tx2gene    = system.file("extdata/tx2gene.rds", package="goiExplorer"),
  outdir     = "salmon_output"
)
}
}
