% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_txt_xlsx}
\alias{write_txt_xlsx}
\title{Export a data.frame to TXT and XLSX}
\usage{
write_txt_xlsx(data, outdir, prefix = base::deparse(base::substitute(data)))
}
\arguments{
\item{data}{A data.frame or tibble to save.}

\item{outdir}{Path to an existing directory.}

\item{prefix}{File name prefix (defaults to the name of \code{data}).}
}
\value{
Invisible \code{NULL}. Files are written to \code{outdir}.
}
\description{
Writes a tab-delimited TXT and an Excel file side-by-side.
}
\details{
This will call \code{utils::write.table()} and \code{writexl::write_xlsx()}.
If \code{prefix} does not exist it will be created based on the data name.
}
\examples{
\dontrun{
df <- matrix(20, 10, 5)
write_txt_xlsx(df, tempdir())
}
}
