% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_agent.R
\name{ai_agent}
\alias{ai_agent}
\title{Chat with your GOI Explorer results}
\usage{
ai_agent(prompt, model = "gpt-4")
}
\arguments{
\item{prompt}{A natural language question about your last pipeline run.}

\item{model}{OpenAI model to use.}
}
\value{
A character response.
}
\description{
Chat with your GOI Explorer results
}
