% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Execute the full RNA-Seq analysis pipeline}
\usage{
run_pipeline(
  input,
  dataType,
  goi = "CYLD",
  parent_outdir,
  abr_healthy = "H",
  abr_case = "D",
  ensemblSpecies = "hsapiens_gene_ensembl",
  rowNamesOfCounts = "external_gene_name",
  interestedTerms = c("cardi", "inflammat"),
  intgroup = c("condition"),
  excludedSample = NA,
  lfcCutoff = 1,
  pCutoff = 0.05,
  pAdjustMethod = "fdr",
  palette = c("#4f8832", "#f79c18"),
  tx2gene = NA
)
}
\arguments{
\item{input}{Character. Path (or list of two) to the directory containing input files:
\itemize{
\item for \code{"counts"}: \code{counts.txt}, \code{gr.txt} or a list of two slots: \code{counts} (table) and \code{gr} (vector)
\item for \code{"salmon"}: \code{ID.txt}, \code{ID.txt}, \code{gr.txt}, plus any Salmon \code{quant.sf} dirs
}}

\item{dataType}{Character, one of \code{"counts"} or \code{"salmon"} indicating
which underlying pipeline to use.}

\item{goi}{a gene of interest}

\item{parent_outdir}{Character. Directory where all outputs (tables, plots) will be written.
If it doesn’t exist, it will be created recursively.}

\item{abr_healthy}{abbreviation for healthy group; for example H for healthy}

\item{abr_case}{abbreviation for case group; for example D for disease}

\item{ensemblSpecies}{e.g. hsapiens_gene_ensembl. You can get that from biomaRt::listDatasets(biomaRt::useEnsembl("ensembl"))$dataset}

\item{rowNamesOfCounts}{the type of the rownames in count table, e.g. external_gene_name}

\item{interestedTerms}{give priority to interested terms (e.g. cardio or inflammation) in otargen::colocalisationsForGene}

\item{intgroup}{the column name of the group, e.g. \code{condition}}

\item{excludedSample}{optional: the samples you want to exclude. should be the names of the corresponded column}

\item{lfcCutoff}{cutoff on logFC}

\item{pCutoff}{cutoff on adj pvalue}

\item{pAdjustMethod}{method to adjusting the pvalues. see ?p.adjust}

\item{palette}{palette to fill the plots}

\item{tx2gene}{Path to an \code{.rds} file or data.frame mapping transcripts -> genes.
a directory also will be created, icluding all outputs.}
}
\value{
A \code{list} with elements:
\itemize{
\item \code{dds}: the \code{DESeqDataSet} object
\item \code{results}: a \code{data.frame} of annotated DE results
\item \code{enrichment}: the \code{enrichResult} object from ORA
}
}
\description{
Dispatches either the “counts” or “salmon” workflow based on \code{dataType},
then runs differential expression, annotation, KEGG, and plots.
}
\examples{
\dontrun{
# counts-based run
meta <- data.frame(
  sample = readLines(system.file("extdata","ID.txt",package="goiExplorer")),
  group  = readLines(system.file("extdata","gr.txt",package="goiExplorer"))
)
run_pipeline(
  input     = system.file("extdata", package = "goiExplorer"),
  dataType  = "counts",
  parent_outdir    = "counts_output/"
)
}
}
